Updated main.py and dataset.py: data visualiser was moved from dataset.py to main.py
